RAG_Practice/LLM_Router/requirements.txt
# Core dependencies
streamlit>=1.31.0
sentence-transformers>=2.3.0
faiss-cpu>=1.9.0

# Additional utilities (installed automatically with sentence-transformers)
# torch and transformers are dependencies of sentence-transformers
